requirements.txt
flask>=1.1.1
flask_SQLAlchemy>=2.4.1
flask_migrate>=2.5.2
requests>=2.22
psycopg2-binary
Flask-WTF
Flask-Bootstrap4>=4.0.0